자폭 몬스터 , 비행 몬스터 추가 , 보스몬스터 컬리전 추가

+++ Directx/Directx/GameEngineContents/Potion_NPC.h
@@ -106,7 +106,7 @@ private:
 	int Item_Number_Check_02 = 0;
 	int Item_Number_Check_03 = 0;
 
-	int Move_X = 0;
-	int Move_Y = 0;
+	size_t Move_X = 0;
+	size_t Move_Y = 0;
 
 };
